Completely refactor file structure
