+++ client/src/models/Deletable.ts
@@ -1,0 +1,3 @@
+export interface Deletable {
+  id: string;
+}
