+++ client/src/models/category.ts
@@ -1,4 +1,4 @@
-import { Deletable } from "../../components/deleteItem.helper";
+import { Deletable } from "./Deletable";
 
 export interface Category extends Deletable {
   id: string;
