+++ client/src/views/EmptyContent.tsx
